requirements.txt
fastapi>=0.109.0,<0.111.0  # Update FastAPI (compatible with Pydantic v2)
uvicorn[standard]>=0.27.0,<0.29.0 # Update Uvicorn

# --- Data Validation ---
pydantic>=2.5.0,<3.0.0 # Use Pydantic v2
email-validator>=2.0.0 # Required for EmailStr in Pydantic v2

# --- LangChain ---
langchain
langchain-core
langchain-community
langchain-neo4j

# --- LLM / Embeddings ---
sentence-transformers>=2.2.2 # Keep or update if needed
torch>=2.0.1 # Keep or update if needed
# huggingface-hub>=0.11.1

# --- Database ---
neo4j>=5.18.0,<6.0.0

# --- Authentication ---
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# --- Utilities ---
python-dotenv>=1.0.0
pypdf>=3.17.0
python-multipart

# --- Frontend ---
jinja2>=3.0.0
aiofiles>=0.8.0